fix: fix errors after merging dev
