+++ package.json
@@ -73,5 +73,8 @@
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript": "5.3.3"
   },
-  "packageManager": "yarn@4.0.2"
+  "packageManager": "yarn@4.0.2",
+  "resolutions": {
+    "@types/react": "^18.0.26"
+  }
 }
