+++ src/app/(main)/my-voyage/[teamId]/ideation/components/IdeationComponentWrapper.tsx
@@ -27,7 +27,7 @@ export async function fetchProjectIdeas({
       `api/v1/voyages/${teamId}/ideations`,
       token,
       "force-cache",
-      CacheTag.ideation
+      CacheTag.ideation,
     );
 
   return await handleAsync(fetchProjectIdeasAsync);
@@ -49,7 +49,7 @@ export default async function IdeationComponentWrapper({
 
   if (user) {
     currentVoyageTeam = user.voyageTeamMembers.find(
-      (voyage) => voyage.voyageTeam.voyage.status.name === "Active"
+      (voyage) => voyage.voyageTeam.voyage.status.name === "Active",
     );
   }
