+++ src/app/(main)/my-voyage/[teamId]/ideation/components/IdeationForm.tsx
@@ -146,7 +146,7 @@ export default function IdeationForm() {
   useEffect(() => {
     if (params.ideationId) {
       const ideation = projectIdeas.find(
-        (project) => project.id === +params.ideationId
+        (project) => project.id === +params.ideationId,
       );
 
       setIdeationData(ideation);
@@ -168,7 +168,7 @@ export default function IdeationForm() {
     () => () => {
       void persistor.purge();
     },
-    []
+    [],
   );
 
   function renderButtonContent() {
