+++ src/app/(main)/my-voyage/[teamId]/sprints/(pages)/add-topic/page.tsx
@@ -1,0 +1,5 @@
+import TopicForm from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/TopicForm";
+
+export default function AddTopicPage() {
+  return <TopicForm />;
+}
