+++ src/app/(main)/my-voyage/[teamId]/sprints/(pages)/create-meeting/page.tsx
@@ -1,0 +1,5 @@
+import MeetingForm from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/MeetingForm";
+
+export default function CreateMeetingPage() {
+  return <MeetingForm />;
+}
