+++ src/app/(main)/my-voyage/[teamId]/sprints/(pages)/edit-meeting/page.tsx
@@ -1,0 +1,5 @@
+import MeetingForm from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/MeetingForm";
+
+export default function EditMeetingPage() {
+  return <MeetingForm />;
+}
