+++ src/app/(main)/my-voyage/[teamId]/sprints/(pages)/voyage-submission/page.tsx
@@ -1,0 +1,5 @@
+import VoyageSubmissionForm from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/VoyageSubmission/VoyageSubmissionForm";
+
+export default function VoyageSubmissionPage() {
+  return <VoyageSubmissionForm />;
+}
