+++ src/app/(main)/my-voyage/[teamId]/sprints/(pages)/weekly-checkin/page.tsx
@@ -1,0 +1,5 @@
+import WeeklyCheckInForm from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/WeeklyCheckIn/WeeklyCheckInForm";
+
+export default function WeeklyCheckInPage() {
+  return <WeeklyCheckInForm />;
+}
