+++ src/app/(main)/my-voyage/[teamId]/sprints/components/forms/VoyageSubmission/VoyageSubmissionForm.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 
 import { scaleOptions } from "./radioGroupDate";
 
-import BaseFormPage from "@/app/(main)/my-voyage/sprints/components/forms/BaseFormPage";
+import BaseFormPage from "@/app/(main)/my-voyage/[teamId]/sprints/components/forms/BaseFormPage";
 import Label from "@/components/inputs/Label";
 import Button from "@/components/Button";
 import Textarea from "@/components/inputs/Textarea";
@@ -16,7 +16,7 @@ import { validateTextInput } from "@/helpers/form/validateInput";
 import TextInput from "@/components/inputs/TextInput";
 
 import { useAppDispatch } from "@/store/hooks";
-import { onOpen } from "@/store/features/modal/modalSlice";
+import { onOpenModal } from "@/store/features/modal/modalSlice";
 
 const validationSchema = z.object({
   projectName: validateTextInput({
@@ -66,7 +66,7 @@ export default function VoyageSubmissionForm() {
 
   const onSubmit: SubmitHandler<ValidationSchema> = (data) => {
     console.log(data);
-    dispatch(onOpen({ type: "voyageSuccess" }));
+    dispatch(onOpenModal({ type: "voyageSuccess", content: "" }));
   };
 
   return (
