--- src/app/(main)/my-voyage/sprints/(pages)/add-topic/page.tsx
@@ -1,5 +1,0 @@
-import TopicForm from "@/app/(main)/my-voyage/sprints/components/forms/TopicForm";
-
-export default function AddTopicPage() {
-  return <TopicForm />;
-}
