--- src/app/(main)/my-voyage/sprints/(pages)/create-meeting/page.tsx
@@ -1,5 +1,0 @@
-import MeetingForm from "@/app/(main)/my-voyage/sprints/components/forms/MeetingForm";
-
-export default function CreateMeetingPage() {
-  return <MeetingForm />;
-}
