--- src/app/(main)/my-voyage/sprints/(pages)/edit-meeting/page.tsx
@@ -1,5 +1,0 @@
-import MeetingForm from "@/app/(main)/my-voyage/sprints/components/forms/MeetingForm";
-
-export default function EditMeetingPage() {
-  return <MeetingForm />;
-}
