--- src/app/(main)/my-voyage/sprints/(pages)/voyage-submission/page.tsx
@@ -1,5 +1,0 @@
-import VoyageSubmissionForm from "@/app/(main)/my-voyage/sprints/components/forms/VoyageSubmission/VoyageSubmissionForm";
-
-export default function VoyageSubmissionPage() {
-  return <VoyageSubmissionForm />;
-}
