--- src/app/(main)/my-voyage/sprints/(pages)/weekly-checkin/page.tsx
@@ -1,5 +1,0 @@
-import WeeklyCheckInForm from "@/app/(main)/my-voyage/sprints/components/forms/WeeklyCheckIn/WeeklyCheckInForm";
-
-export default function WeeklyCheckInPage() {
-  return <WeeklyCheckInForm />;
-}
