+++ src/components/modals/Example2Modal.tsx
@@ -69,7 +69,7 @@ export default function Example2Modal() {
       onOpen({
         context: "success",
         message: "Your information has been updated",
-      })
+      }),
     );
     handleClose();
   };
