+++ src/utils/handleAsync.ts
@@ -4,7 +4,7 @@ type AsyncFunction<X> = () => Promise<X>;
 export type AsyncActionResponse<X> = [X | null, AppError | null];
 
 export async function handleAsync<X>(
-  asyncFn: AsyncFunction<X>
+  asyncFn: AsyncFunction<X>,
 ): Promise<AsyncActionResponse<X>> {
   try {
     const result = await asyncFn();
